Included a smart drive adjustment when an enemy has been detected. Set some fine tuning on the sensor monitoring.

+++ EV3Brick/State/Normal/NormalIdentifyImpl.cs
@@ -8,7 +8,7 @@ namespace SPSGrp1Grp2.Cunt.State.Normal
         private const int Level = EventQueue.StateLevel.Level2;
 
         internal NormalIdentifyImpl() {
-            StateTimer.TickTimeout = Ev3Timer.TickTime.Long;
+            StateTimer.TickTimeout = Ev3Timer.TickTime.Short;
         }
 
         protected override void PerformRecurrentAction()
